pyproject.toml
[project]
name = "celine-ontologies"
version = "0.1.0"
description = "CELINE Ontologies support libraries"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "rdflib>=7.5.0",
    "typer>=0.21.1",
    "pydantic>=2.0",
    "pyshacl>=0.26.0",
    "jsonschema>=4.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.uv]  # optional, helps uv handle locking
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["src/celine/mapper", "src/celine/ontologies"]
sources = {"src" = ""}

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = true
check_untyped_defs = true


[project.scripts]
celine-ontologies = "celine.ontologies.main:run"


[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
update_pyproject_toml = true 
commit_parser = "conventional"

build_command = """
    uv lock --upgrade-package "$PACKAGE_NAME"
    git add uv.lock
    uv build
"""

[tool.semantic_release.commit_parser_options]
minor_tags = ["feat"]
patch_tags = ["fix", "perf", "chore"]
parse_squash_commits = true
ignore_merge_commits = true

